+++ packages/identity-integration/README.md
@@ -3,7 +3,7 @@
 ## BREAKING CHANGE 0.0.16
 
 - Необходим глобальный провайдер `KratosClientProvider`. По умолчанию он предоставляет
-  стандартный клиент `kratos`, однако его можно будет изменять/расширять по потребности в 
+  стандартный клиент `kratos`, однако его можно будет изменять/расширять по потребности в
   проекте, а также URL редиректа для некоторых ошибок.
 - Добавлена кастомизация URL для редиректа в случае ошибок, например пользователь уже авторизован.
 - Расширен конструктов класса `KratosClient`
