+++ packages/identity-integration/package.json
@@ -15,6 +15,7 @@
     "@ory/client": "1.2.11",
     "@ory/integrations": "1.1.5",
     "@ory/kratos-client": "1.0.0",
+    "@types/tldjs": "2.3.4",
     "tldjs": "2.3.1"
   },
   "devDependencies": {
